Update 20180320
Mainnet: Delete the admin.macaroon, readonly.macaroon and macaroons.db files before the switch to mainnet. This is always needed before a new wallet is created.

Bitcoin and Lightning: Copy credentials for bitcoind and lnd to user "admin" home directory. As this is a superuser anyway there's no sense in always switching to the user "bitcoin" session.