core wallet needs to be enabled to run eps #133 (#137)
* core wallet needs to be enabled to run eps # 133

Solution to issue # 133 by enabling the bitcoin core wallet. As commenting out `disablewallet=1` has not worked, the value is changed in this proposal.

* Streamline config update

The bitcoind service does not need to be stopped before editing the config file. A restart of the RaspiBolt should not be necessary, just restarting the daemon afterwards is enough.