Fix for post LND 0.5 break
The macaroon location changed and is unreadable by the "admin" user and therefore creates an issue where the LND information is not displayed. The lncli option default the macaroon location as well as the tls cert location so those options are not needed. By removing those options this script is able to be run by admin and bitcoin. Root, while able to run this script, does not have the macaroon copied and therefore will not get the LND information.