Updated auto unlock script for LND 0.5
Added and changed some lines to make the autounlock able to work on LND0.5. I made a pull this pull request on the basis of the proposed adaptations of issue#254 by @zwarbo. I also used a piece of code from the pull request of issue#246 submitted by @martinatime to check for testnet/mainnet.

(I also added the contributers of this commit to the file, as the original writers were mentioned aswell. Not sure if this is ok. )