Electrs: make integral part of RaspiBolt (WIP)
Despite Bitcoin full nodes being en vogue lately, this is mostly due to
people using the Lightning Network.

The importance of validating ones own on-chain transactions,
especially for important transactions, that are typically handled with
hardware wallets, is not obvious to many.

Having a full Electrum server to use (and maybe share with friends and
family) is an important addition to the RaspiBolt.

This commit:
* adds the Electrum part, consisting of
  * Electrs installation
  * using NGINX as reverse proxy
  * adding a Tor hidden service for remote access
* describes how to connect Electrum and BitBoxApp
  to your own full node