Delete raspibolt_68_shango.md
The Shango app does not seem to be developped/maintained anymore. The last commit in the Github repo was in June 2019, so almost 2.5 years ago... Also, the website shangoapp.com is now dead and does not redirect to anything.

For security reason, it might be better to remove this installation guide.