Delete raspibolt_64_electrum.md
This guide seems outdated and seems a bit redundant as the main guide already contains the installation of Electrs.
I remember as a beginner to have been confused by this bonus guide, not understanding how it related to Electrs that I had already installed.

The last section explaining how to connect Electrum Desktop to one server only (our Electrs) is very useful and missing in the main guide. So this part could be salvaged and added at the end of the Electrs page.