Lightning: general update, incl watchtowers & RTL
All Bitcoin Lightning guides are now part of the 'Lightning' section,
with these individual subsections:

- LND
- Ride The Lightning

The LND guide is updated to use the latest versions and ARM64 builds.
Usage of watchtowers as a client, as well as default unlock of the
LND wallet through a password file is added.

To make node management easier, RTL now exposes full admin
capabilities with a web interface.