bonus: boot from microsd card
The new RaspiBolt version 3 assumes that all external drives can be
used to boot the Raspberry Pi.

This seems to hold true for most modern drives, but if this does not
work with a specific drive, the reader is stuck without any fallback
instructions.

This change adds a bonus guide on how to boot from a microSD card
and mount an external drive to /data/ to use for data storage. This
method should work with the broadest range of devices.