Tor: move web endpoints to TLS/SSL
While the Tor network is encrypted in itself, malicious actors can run
nodes that can snoop on traffic. A password transmitted in cleartext
could allow to log in to RTL and steal your Bitcoin (if no 2FA enabled).

This change moves all web endpoint exposed as Tor .onion addresses
to the NGINX ports, adding TLS/SSL for additional encryption on top of
the Tor network.

The downside is that this will show a browser warning due to the
self-signed certificate.