BTC RPC Explorer: update to v3.3.0
The latest release brings a ton of new features and is well worth
the update.

Also, the current update guide has some issues that need to be
addressed. The NPM install process changes files in the repository
that block the checkout of a newer branch later on. The repo
needs to be reset first, and local changes discarded.

Additionally, v3.3.0 was tagged without a date, not sure how
that's even possible. But this breaks the 'git describe' command,
so at least temporarily we fall back to 'git tag'.