bitcoind: use 'startupnotify' for group perms (#846)
Bitcoin Core is accessible to applications run by different users
through the authentication cookie. Users that are a member of the
group "bitcoin" need read-only access to /data/bitcoin/.cookie and
can then authenticate themselves without username and password.

By default, Bitcoin Core does not create the .cookie file with
group permissions. As of now, the bitcoind argument `sysperms=1`
is used with UMask=0027 to override this. Bitcoind only allows
this, if `disablewallet=1` is set. This breaks compatibility with
applications that rely on the Bitcoin Core wallet like JoinMarket
or Electrum Personal Server.

This change ensures the same behavior (making the .cookie file
group accessible) with a different method: executing a shell
command after startup of bitcoind, configured with the argument
`startupnotify` and triggered by bitcoind itself.