welcome script: update to v3 and improve syntax (#850)
* Monitoring channel.db size in welcome/motd script (#809)

* Add channel.db size to Welcome script

  Monitoring the channel.db size is critical as a very large DB can lead to node crashes (and potentially DB corruption & SCB recovery) on low-resource devices like Raspberry Pis (e.g. https://github.com/lightningnetwork/lnd/issues/5705). 
This update adds a line to the LND section of the welcome script to show the channel.db size

  For now the size is obtained by: channeldb=$(du -h /mnt/ext/lnd/data/graph/mainnet/channel.db | head -c4)
It would be better to get a numerical value and then add some simple logical tests to colour the size green, orange or red when it reaches a certain large size (e.g. 500MN orange, 1GB red). What would be the best way to do this? (i.e. get a numerical value for the db size rather than a string like now)

* updated variable name

  I changed the variable name from 'channeldb' to channel_db_size' to make it more meaningful.

* Update 20-raspibolt-welcome

  Replaced "head -c4" with "awk '{print $1}'" as advised by Pablo [here](https://t.me/raspibolt/3521) as a most robust way of getting the relevant section of the output (thanks to him!)

* Changed title to 'Channel.db size:'

* update MOTD welcome script for v3

The MOTD welcome script grew significantly over time and does
not follow best shell script practices.

This change updates it to reflect the setup of RaspiBolt v3 and
improves overall script quality.

* system overview: trap Ctrl-C for clean exit

When the user presses Ctrl-C, the color encoding might disrupt the
regular command line.

This change traps the Ctrl-C abort commands SIGINT and SIGTERM,
and exits the script cleanly.

* system overview: address pr feedback

Co-authored-by: Indra <86242283+VajraOfIndra@users.noreply.github.com>