Update privacy.md
When following the RaspiBolt guide to (optionally) SSH in to the raspibolt via Tor I hit a problem when following the advice v3 guide pointed to here:
https://gitlab.torproject.org/legacy/trac/-/wikis/doc/TorifyHOWTO/Putty
Because I was running Tor on windows, I think Tor Bundle port is now 9150 rather than 9050?
I think this is just the tor browser-bundle specific. Tor itself still listens on 9050
But the advice on the linked to page was to connect to 9050.