Adding the lnd stop command in systemd unit file
Without this command, an incorrect stop occurs - systemd sends a kill signal, which is not processed correctly by LND - part of the services inside LND stops, but the process itself continues to work. Then systemd sends signal `9`, which kills the process. In this case, data loss (for example, channels) may occur, which will lead to the loss of funds.