Improve the Upgrade section of the regolancer bonus guide (#1162)
* Improve the Upgrade section

Changed the upgrade command from a specific version to the latest one. This makes it easier for users to upgrade, since they don't need to check the exact version number of the latest release.

* Change the main install command to @latest

* add GOARCH=arm64

Add GOARCH=arm64 to the installation command

* Remove a note that is no longer required

Remove "Note: Adjust the "regolancer@v1.6.3" part from the commands below to the actual version you have installed.". We no longer need this info since we are installing @latest.

* Changing the path to the sample conf file...

...and adding the note that this might be adjusted, depending on the actual version that was just installed.