Increase default mempool to 2 GB
Recently we have seen many node operators struggling with transactions that are't broadcasted due to being below the purging limit. If a (force) close transactions fee rate is below the purging limit, the node will drop it immediately and therefore it won't be broadcasted at all. We should increase the default mempool size significantly to mitigate this issue.