Provide a Nextcloud Backup option for lightning channels
The raspiblitz has an option to backupt the channel.backup to a self-hosted (or any) nextcloud instance. 
As I have one, this was a feature I liked.
I implemented a section to the provided bash script which is triggered by inotify which allows making a backup to nextcloud with a specified webdav URL, username and password.

The guide itself was not altered as I just want to propose this for the main branch as an idea.